Refactor : entity.DeliveryStatus -> dto.DeliveryStatusDto
