+++ order-server/src/main/java/com/inbobwetrust/config/swaggerdoc/DeliveryControllerSwaggerDoc.java
@@ -4,7 +4,7 @@ import com.inbobwetrust.aop.ApiResult;
 import com.inbobwetrust.model.dto.DeliveryCreateDto;
 import com.inbobwetrust.model.dto.DeliverySetRiderDto;
 import com.inbobwetrust.model.entity.Delivery;
-import com.inbobwetrust.model.entity.DeliveryStatus;
+import com.inbobwetrust.model.dto.DeliveryStatusDto;
 
 import io.swagger.annotations.*;
 
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 public interface DeliveryControllerSwaggerDoc {
     @ApiOperation(value = "주문상태확인", notes = "주문번호에 해당하는 주문상태 반환")
     @ApiResponses(value = {@ApiResponse(code = 200, message = "성공", response = ApiResult.class)})
-    DeliveryStatus getDeliveryStatus(Long orderId);
+    DeliveryStatusDto getDeliveryStatus(Long orderId);
 
     @ApiOperation(value = "픽업완료", notes = "주문의 상태를 픽업완료로 갱신")
     @ApiResponses(value = {@ApiResponse(code = 200, message = "성공", response = ApiResult.class)})
