+++ order-server/src/main/java/com/inbobwetrust/model/dto/DeliveryStatusDto.java
@@ -1,17 +1,18 @@
-package com.inbobwetrust.model.entity;
+package com.inbobwetrust.model.dto;
 
+import com.inbobwetrust.model.entity.OrderStatus;
 import lombok.*;
 
 @Getter
 @ToString
 @NoArgsConstructor
 @EqualsAndHashCode
-public class DeliveryStatus {
+public class DeliveryStatusDto {
     private Long orderId;
     private OrderStatus orderStatus;
 
     @Builder
-    public DeliveryStatus(Long orderId, OrderStatus orderStatus) {
+    public DeliveryStatusDto(Long orderId, OrderStatus orderStatus) {
         this.orderId = orderId;
         this.orderStatus = orderStatus;
     }
