+++ order-server/src/main/java/com/inbobwetrust/repository/DeliveryRepository.java
@@ -1,7 +1,7 @@
 package com.inbobwetrust.repository;
 
 import com.inbobwetrust.model.entity.Delivery;
-import com.inbobwetrust.model.entity.DeliveryStatus;
+import com.inbobwetrust.model.dto.DeliveryStatusDto;
 
 import java.util.List;
 import java.util.Optional;
@@ -15,7 +15,7 @@ public interface DeliveryRepository {
 
     boolean update(Delivery delivery);
 
-    Optional<DeliveryStatus> findDeliveryStatusByOrderId(Long orderId);
+    Optional<DeliveryStatusDto> findDeliveryStatusByOrderId(Long orderId);
 
     Optional<Delivery> findById(Long orderId);
 }
